ScriptRunner UserInterface separate class
