+++ app/src/main/java/com/mathewsachin/fategrandautomata/accessibility/ScriptRunnerService.kt
@@ -1,7 +1,6 @@
 package com.mathewsachin.fategrandautomata.accessibility
 
 import android.accessibilityservice.AccessibilityService
-import android.annotation.SuppressLint
 import android.app.Activity.RESULT_OK
 import android.app.AlarmManager
 import android.app.NotificationChannel
@@ -9,23 +8,15 @@ import android.app.NotificationManager
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
-import android.graphics.Color
-import android.graphics.PixelFormat
 import android.media.projection.MediaProjection
 import android.media.projection.MediaProjectionManager
 import android.os.Build
 import android.os.SystemClock
-import android.util.DisplayMetrics
-import android.view.*
 import android.view.accessibility.AccessibilityEvent
-import android.widget.FrameLayout
 import android.widget.ImageButton
 import android.widget.Toast
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
-import androidx.core.graphics.BlendModeColorFilterCompat
-import androidx.core.graphics.BlendModeCompat
-import androidx.core.view.postDelayed
 import com.mathewsachin.fategrandautomata.R
 import com.mathewsachin.fategrandautomata.core.*
 import com.mathewsachin.fategrandautomata.imaging.MediaProjectionRecording
@@ -42,23 +33,17 @@ import com.mathewsachin.fategrandautomata.scripts.entrypoints.SupportImageMaker
 import com.mathewsachin.fategrandautomata.scripts.enums.ScriptModeEnum
 import com.mathewsachin.fategrandautomata.scripts.prefs.Preferences
 import com.mathewsachin.fategrandautomata.ui.MainActivity
-import com.mathewsachin.fategrandautomata.ui.highlightView
 import com.mathewsachin.fategrandautomata.ui.support_img_namer.SupportImageIdKey
 import com.mathewsachin.fategrandautomata.ui.support_img_namer.SupportImageNamerActivity
 import com.mathewsachin.fategrandautomata.util.AndroidImpl
-import kotlin.math.roundToInt
+import kotlin.time.seconds
 
 class ScriptRunnerService : AccessibilityService() {
     companion object {
         var Instance: ScriptRunnerService? = null
     }
 
-    private lateinit var layout: FrameLayout
-    private lateinit var windowManager: WindowManager
-    private lateinit var layoutParams: WindowManager.LayoutParams
-    private lateinit var highlightLayoutParams: WindowManager.LayoutParams
-    private lateinit var scriptCtrlBtn: ImageButton
-    private val metrics = DisplayMetrics()
+    private lateinit var userInterface: ScriptRunnerUserInterface
     private var sshotService: IScreenshotService? = null
     private var gestureService: IGestureService? = null
     private var superUser: SuperUser? = null
@@ -110,8 +95,7 @@ class ScriptRunnerService : AccessibilityService() {
             return false
         }
 
-        windowManager.addView(highlightView, highlightLayoutParams)
-        windowManager.addView(layout, layoutParams)
+        userInterface.show()
 
         serviceStarted = true
 
@@ -123,7 +107,7 @@ class ScriptRunnerService : AccessibilityService() {
             if (MediaProjectionToken != null) {
                 mediaProjection =
                     mediaProjectionManager.getMediaProjection(RESULT_OK, MediaProjectionToken)
-                MediaProjectionScreenshotService(mediaProjection!!, metrics)
+                MediaProjectionScreenshotService(mediaProjection!!, userInterface.metrics)
             } else RootScreenshotService(
                 getSuperUser()
             )
@@ -169,8 +153,7 @@ class ScriptRunnerService : AccessibilityService() {
         ScreenshotManager.releaseMemory()
         clearImageCache()
 
-        windowManager.removeView(layout)
-        windowManager.removeView(highlightView)
+        userInterface.hide()
         serviceStarted = false
 
         hideForegroundNotification()
@@ -181,9 +164,7 @@ class ScriptRunnerService : AccessibilityService() {
     private var entryPoint: EntryPoint? = null
 
     private fun onScriptExit() {
-        scriptCtrlBtn.post {
-            scriptCtrlBtn.setImageResource(R.drawable.ic_play)
-        }
+        userInterface.setPlayIcon()
 
         clearSupportCache()
 
@@ -193,7 +174,7 @@ class ScriptRunnerService : AccessibilityService() {
         val rec = recording
         if (rec != null) {
             // record for 2 seconds more to show things like error messages
-            scriptCtrlBtn.postDelayed(2000) { rec.close() }
+            userInterface.postDelayed(2.seconds) { rec.close() }
         }
 
         recording = null
@@ -220,7 +201,7 @@ class ScriptRunnerService : AccessibilityService() {
         }
 
         if (Preferences.RecordScreen && mediaProjection != null) {
-            recording = MediaProjectionRecording(mediaProjection!!, metrics)
+            recording = MediaProjectionRecording(mediaProjection!!, userInterface.metrics)
         }
 
         // Reset the value just in case it wasn't already
@@ -229,10 +210,9 @@ class ScriptRunnerService : AccessibilityService() {
         entryPoint = getEntryPoint().apply {
             scriptExitListener = ::onScriptExit
 
-            scriptCtrlBtn.setImageResource(R.drawable.ic_stop)
+            userInterface.setStopIcon()
             if (recording != null) {
-                scriptCtrlBtn.drawable.colorFilter = BlendModeColorFilterCompat
-                    .createBlendModeColorFilterCompat(Color.RED, BlendModeCompat.SRC_ATOP)
+                userInterface.showAsRecording()
             }
 
             run()
@@ -280,116 +260,24 @@ class ScriptRunnerService : AccessibilityService() {
         super.onTaskRemoved(rootIntent)
     }
 
-    @SuppressLint("RtlHardcoded")
+    fun registerScriptCtrlBtnListeners(scriptCtrlBtn: ImageButton) {
+        scriptCtrlBtn.setOnClickListener {
+            if (scriptStarted) {
+                stopScript()
+            } else startScript()
+        }
+    }
+
     override fun onServiceConnected() {
         Instance = this
         AutomataApi.registerPlatform(AndroidImpl(this))
 
-        windowManager = getSystemService(Context.WINDOW_SERVICE) as WindowManager
-
-        layout = FrameLayout(this)
-        layoutParams = WindowManager.LayoutParams().apply {
-            type = WindowManager.LayoutParams.TYPE_ACCESSIBILITY_OVERLAY
-            format = PixelFormat.TRANSLUCENT
-            flags = WindowManager.LayoutParams.FLAG_NOT_FOCUSABLE
-            width = WindowManager.LayoutParams.WRAP_CONTENT
-            height = WindowManager.LayoutParams.WRAP_CONTENT
-            @SuppressLint("RtlHardcoded")
-            gravity = Gravity.LEFT or Gravity.TOP
-            x = 0
-            y = 0
-        }
-
-        val inflater = LayoutInflater.from(this)
-        inflater.inflate(R.layout.script_runner, layout)
-
-        highlightLayoutParams = WindowManager.LayoutParams().apply {
-            type = WindowManager.LayoutParams.TYPE_ACCESSIBILITY_OVERLAY
-            format = PixelFormat.TRANSLUCENT
-            flags =
-                WindowManager.LayoutParams.FLAG_NOT_FOCUSABLE or WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE
-            width = WindowManager.LayoutParams.MATCH_PARENT
-            height = WindowManager.LayoutParams.MATCH_PARENT
-        }
-
-        scriptCtrlBtn = layout.findViewById<ImageButton>(R.id.script_toggle_btn).also {
-            it.setOnClickListener {
-                if (scriptStarted) {
-                    stopScript()
-                } else startScript()
-            }
-
-            it.setOnTouchListener(::scriptCtrlBtnOnTouch)
-        }
-
         mediaProjectionManager =
             getSystemService(Context.MEDIA_PROJECTION_SERVICE) as MediaProjectionManager
 
-        windowManager.defaultDisplay.getRealMetrics(metrics)
-
-        // Retrieve images in Landscape
-        if (metrics.heightPixels > metrics.widthPixels) {
-            metrics.let {
-                val temp = it.widthPixels
-                it.widthPixels = it.heightPixels
-                it.heightPixels = temp
-            }
-        }
-
-        layoutParams.y = metrics.widthPixels
+        userInterface = ScriptRunnerUserInterface(this)
 
         super.onServiceConnected()
-    }
-
-    private fun getMaxBtnCoordinates(): Location {
-        val rotation = windowManager.defaultDisplay.rotation
-        val rotate = rotation == Surface.ROTATION_0 || rotation == Surface.ROTATION_180
-
-        val x = (if (rotate) {
-            metrics.heightPixels
-        } else metrics.widthPixels) - layout.measuredWidth
-        val y = (if (rotate) {
-            metrics.widthPixels
-        } else metrics.heightPixels) - layout.measuredHeight
-
-        return Location(x, y)
-    }
-
-    private val dragStopwatch = Stopwatch()
-    private var dX = 0f
-    private var dY = 0f
-    private var lastAction = 0
-
-    private fun scriptCtrlBtnOnTouch(_View: View, Event: MotionEvent): Boolean {
-        return when (Event.actionMasked) {
-            MotionEvent.ACTION_DOWN -> {
-                val (maxX, maxY) = getMaxBtnCoordinates()
-                dX = layoutParams.x.coerceIn(0, maxX) - Event.rawX
-                dY = layoutParams.y.coerceIn(0, maxY) - Event.rawY
-                lastAction = MotionEvent.ACTION_DOWN
-                dragStopwatch.start()
-
-                false
-            }
-            MotionEvent.ACTION_MOVE -> {
-                val newX = Event.rawX + dX
-                val newY = Event.rawY + dY
-
-                if (dragStopwatch.elapsedMs > ViewConfiguration.getLongPressTimeout()) {
-                    val (mX, mY) = getMaxBtnCoordinates()
-                    layoutParams.x = newX.roundToInt().coerceIn(0, mX)
-                    layoutParams.y = newY.roundToInt().coerceIn(0, mY)
-
-                    windowManager.updateViewLayout(layout, layoutParams)
-
-                    lastAction = MotionEvent.ACTION_MOVE
-                }
-
-                true
-            }
-            MotionEvent.ACTION_UP -> lastAction == MotionEvent.ACTION_MOVE
-            else -> false
-        }
     }
 
     override fun onInterrupt() {}
